normalised target cost weights so that they sum to one; normalised join cost weights so that they sum to one; added a relative weight factor for target and join costs; enhanced the join cost function to be able to take into account phonetic features from the target.
git-svn-id: https://mary.opendfki.de/repos/trunk@391 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/mary/unitselection/voiceimport/JoinCostFileMaker.java
@@ -52,7 +52,7 @@ public class JoinCostFileMaker implements VoiceImportComponent {
     private BasenameList bnl = null;
     
     private int numberOfFeatures = 0;
-    private Float[] fw = null;
+    private float[] fw = null;
     private String[] wfun = null;
     
     /** Constructor */
@@ -105,14 +105,14 @@ public class JoinCostFileMaker implements VoiceImportComponent {
         /****************************/
         /* Load the weight vectors */
         Object[] weightData = JoinCostFeatures.readJoinCostWeightsFile( db.joinCostWeightsFileName() );
-        fw = (Float[]) weightData[0];
+        fw = (float[]) weightData[0];
         wfun = (String[]) weightData[1];
         numberOfFeatures = fw.length;
         /* Output those vectors */
         try {
             jcf.writeInt( fw.length );
             for ( int i = 0; i < fw.length; i++ ) {
-                jcf.writeFloat( fw[i].floatValue() );
+                jcf.writeFloat( fw[i] );
                 jcf.writeUTF( wfun[i] );
             }
         }
